# api/openapi.yaml
openapi: 3.1.0
info:
  title: SMS Gateway API
  version: "1.0"
  description: |
    Send SMS messages, manage credits, and read delivery reports.

    Every request is authenticated with an API key sent as a bearer token.
    Message submissions are rate limited per customer; see the
    `X-RateLimit-*` and `Retry-After` headers. Use `client_ref` to make
    submissions and charges safe to retry.
servers:
  - url: /
security:
  - apiKey: []

tags:
  - name: Messages
  - name: Balance
  - name: Reports

paths:
  /v1/messages:
    post:
      tags: [Messages]
      summary: Send one message
      operationId: sendMessage
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendMessageRequest" }
            example: { to: "+989121234567", text: "Your code is 482913", type: normal, client_ref: otp-1 }
      responses:
        "202":
          description: Accepted, stored, paid for, and queued; or an idempotent replay.
          headers:
            Idempotent-Replayed: { $ref: "#/components/headers/IdempotentReplayed" }
            X-RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/InsufficientCredits" }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
    get:
      tags: [Messages]
      summary: List messages, newest first
      operationId: listMessages
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/Status" } }
        - { name: type, in: query, schema: { $ref: "#/components/schemas/Type" } }
        - { name: recipient, in: query, schema: { type: string, example: "+989121234567" } }
        - { $ref: "#/components/parameters/Since" }
        - { $ref: "#/components/parameters/Until" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: A page of messages.
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Message" } }
                  next_cursor: { type: [string, "null"] }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  /v1/messages/batch:
    post:
      tags: [Messages]
      summary: Send up to 500 messages atomically
      description: |
        Either every message is accepted and the total cost is debited, or
        nothing is recorded. A batch is an idempotent replay only if every
        item has a `client_ref` and all of them match earlier messages
        exactly; any other overlap is a conflict.
      operationId: sendBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              additionalProperties: false
              properties:
                messages:
                  type: array
                  minItems: 1
                  maxItems: 500
                  items: { $ref: "#/components/schemas/SendMessageRequest" }
      responses:
        "202":
          description: All messages accepted, in request order; or a full replay.
          headers:
            Idempotent-Replayed: { $ref: "#/components/headers/IdempotentReplayed" }
            X-RateLimit-Limit: { $ref: "#/components/headers/RateLimitLimit" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/RateLimitRemaining" }
          content:
            application/json:
              schema:
                type: object
                required: [messages, total_cost]
                properties:
                  messages: { type: array, items: { $ref: "#/components/schemas/Message" } }
                  total_cost: { type: integer, format: int64 }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "402": { $ref: "#/components/responses/InsufficientCredits" }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }

  /v1/messages/{id}:
    get:
      tags: [Messages]
      summary: Get one message
      operationId: getMessage
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The message.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  /v1/balance:
    get:
      tags: [Balance]
      summary: Get the current balance
      operationId: getBalance
      responses:
        "200":
          description: The balance in credits.
          content:
            application/json:
              schema:
                type: object
                required: [balance, updated_at]
                properties:
                  balance: { type: integer, format: int64, example: 9412 }
                  updated_at: { $ref: "#/components/schemas/Timestamp" }
        "401": { $ref: "#/components/responses/Unauthorized" }

  /v1/balance/charges:
    post:
      tags: [Balance]
      summary: Add credits
      description: Adds credits idempotently by `client_ref`. Represents a completed top-up.
      operationId: charge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, client_ref]
              additionalProperties: false
              properties:
                amount: { type: integer, format: int64, minimum: 1, maximum: 1000000000 }
                client_ref: { $ref: "#/components/schemas/ClientRef" }
            example: { amount: 5000, client_ref: topup-2026-09-21-01 }
      responses:
        "201":
          description: Credits added, or an idempotent replay with the current balance.
          headers:
            Idempotent-Replayed: { $ref: "#/components/headers/IdempotentReplayed" }
          content:
            application/json:
              schema:
                type: object
                required: [transaction, balance]
                properties:
                  transaction: { $ref: "#/components/schemas/Transaction" }
                  balance: { type: integer, format: int64 }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/IdempotencyConflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  /v1/balance/transactions:
    get:
      tags: [Balance]
      summary: List credit transactions, newest first
      operationId: listTransactions
      parameters:
        - { name: kind, in: query, schema: { type: string, enum: [charge, debit, refund] } }
        - { $ref: "#/components/parameters/Since" }
        - { $ref: "#/components/parameters/Until" }
        - { $ref: "#/components/parameters/Limit" }
        - { $ref: "#/components/parameters/Cursor" }
      responses:
        "200":
          description: A page of transactions.
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Transaction" } }
                  next_cursor: { type: [string, "null"] }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

  /v1/reports/summary:
    get:
      tags: [Reports]
      summary: Aggregate statistics for messages accepted in a time range
      operationId: summary
      parameters:
        - name: since
          in: query
          description: Start of the range; defaults to 24 hours before `until`.
          schema: { type: string, format: date-time }
        - name: until
          in: query
          description: End of the range (exclusive); defaults to now. The range may not exceed 31 days.
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: The summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Summary" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/ValidationFailed" }

components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: "API key, for example `Authorization: Bearer sk_3Qm8x...`"

  parameters:
    Since:
      name: since
      in: query
      description: Only items created at or after this time.
      schema: { type: string, format: date-time }
    Until:
      name: until
      in: query
      description: Only items created before this time.
      schema: { type: string, format: date-time }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      description: The `next_cursor` of the previous page.
      schema: { type: string }

  headers:
    IdempotentReplayed:
      description: Present with value `true` when the response repeats an earlier result.
      schema: { type: string, enum: ["true"] }
    RateLimitLimit:
      description: The customer's rate limit in messages per second.
      schema: { type: integer }
    RateLimitRemaining:
      description: Messages the rate limiter admits right now.
      schema: { type: integer }

  schemas:
    Timestamp:
      type: string
      format: date-time
      example: "2026-09-21T10:15:30.123Z"
    ClientRef:
      type: string
      pattern: "^[A-Za-z0-9._:-]{1,64}$"
      description: Customer reference that makes the request idempotent; unique per customer.
    Type:
      type: string
      enum: [normal, express]
    Status:
      type: string
      enum: [accepted, sent, delivered, undelivered, failed, expired]
    SendMessageRequest:
      type: object
      required: [to, text]
      additionalProperties: false
      properties:
        to:
          type: string
          pattern: "^\\+[1-9][0-9]{7,14}$"
          description: Recipient in E.164 format.
        text:
          type: string
          description: Message text; at most 10 segments (1,530 GSM-7 septets or 670 UCS-2 units).
        type:
          allOf: [{ $ref: "#/components/schemas/Type" }]
          default: normal
        client_ref: { $ref: "#/components/schemas/ClientRef" }
    Message:
      type: object
      required: [id, type, to, text, encoding, segments, cost, status, attempts, failure_reason,
                 sla_breached, client_ref, accepted_at, sent_at, completed_at]
      properties:
        id: { type: string, format: uuid }
        type: { $ref: "#/components/schemas/Type" }
        to: { type: string }
        text: { type: string }
        encoding: { type: string, enum: [gsm7, ucs2] }
        segments: { type: integer, minimum: 1, maximum: 10 }
        cost: { type: integer, format: int64, description: Credits charged. }
        status: { $ref: "#/components/schemas/Status" }
        attempts: { type: integer }
        failure_reason:
          type: [string, "null"]
          enum: [rejected, attempts_exhausted, expired, null]
        sla_breached: { type: boolean, description: Always false for normal messages. }
        client_ref: { type: [string, "null"] }
        accepted_at: { $ref: "#/components/schemas/Timestamp" }
        sent_at: { type: [string, "null"], format: date-time }
        completed_at: { type: [string, "null"], format: date-time }
    Transaction:
      type: object
      required: [id, kind, amount, message_id, client_ref, created_at]
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [charge, debit, refund] }
        amount: { type: integer, format: int64, description: Positive for charges and refunds, negative for debits. }
        message_id: { type: [string, "null"], format: uuid }
        client_ref: { type: [string, "null"] }
        created_at: { $ref: "#/components/schemas/Timestamp" }
    Summary:
      type: object
      required: [since, until, totals, by_status, by_type]
      properties:
        since: { $ref: "#/components/schemas/Timestamp" }
        until: { $ref: "#/components/schemas/Timestamp" }
        totals:
          type: object
          required: [messages, segments, credits_spent, credits_refunded]
          properties:
            messages: { type: integer, format: int64 }
            segments: { type: integer, format: int64 }
            credits_spent: { type: integer, format: int64 }
            credits_refunded: { type: integer, format: int64 }
        by_status:
          type: object
          additionalProperties: { type: integer, format: int64 }
        by_type:
          type: object
          properties:
            normal:
              type: object
              properties:
                messages: { type: integer, format: int64 }
                credits: { type: integer, format: int64 }
            express:
              type: object
              properties:
                messages: { type: integer, format: int64 }
                credits: { type: integer, format: int64 }
                sla_breached: { type: integer, format: int64 }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, request_id]
          properties:
            code: { type: string, description: Stable machine-readable code. }
            message: { type: string }
            details:
              type: [array, "null"]
              items:
                type: object
                required: [field, code, message]
                properties:
                  field: { type: string, example: "messages[2].to" }
                  code:
                    type: string
                    enum: [required, invalid_format, out_of_range, too_many_items, duplicate,
                           text_too_long, invalid_text, conflict]
                  message: { type: string }
            request_id: { type: string }

  responses:
    InvalidJSON:
      description: The body is not valid JSON or contains unknown fields.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing or invalid API key.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    InsufficientCredits:
      description: The balance does not cover the cost; nothing was recorded.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: The resource does not exist or belongs to another customer.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    IdempotencyConflict:
      description: A client_ref was already used for a different request.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    ValidationFailed:
      description: One or more fields are invalid; see details.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    RateLimited:
      description: The message rate limit was exceeded.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }